requirements.txt
pyaudio>=0.2.11
numpy>=1.21.0
pyFluidSynth>=1.3.0

